% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{kpi}
\alias{kpi}
\title{kpi}
\usage{
kpi(awakening_table, swarming_table, maturing_table, hsi_table, topography)
}
\arguments{
\item{awakening_table}{The data.frame returned by the awakening() function.}

\item{swarming_table}{The data.frame returned by the swarming() function.}

\item{maturing_table}{The data.frame returned by the maturing() function.}

\item{hsi_table}{The data.frame returned by the hsc() function.}

\item{topography}{The raster stack returned by the topo_comp() function.}
}
\value{
A raster stack containing the interpolated phenological indicators and hydraulic stress index, restricted to pure spruce forests.
}
\description{
This function spatializes phenological indicators (`awakening_doy`, `swarming_doy`, `maturing_doy`) and hydraulic stress index (`hsi`) using IDW or ordinary kriging with an auto-fitted exponential model, depending on the number of DRIAS points present in the study area.
}
\examples{
\dontrun{
 spat_ind <- kpi(awakening_table, swarming_table, maturing_table, hsi_table, topography)
}
}
