% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{maturing}
\alias{maturing}
\title{maturing}
\usage{
maturing(drias_table, swarming_table, topography)
}
\arguments{
\item{drias_table}{The DRIAS table processed by the phloem_rm() function.}

\item{swarming_table}{The data.frame returned by the swarming() function.}

\item{topography}{The raster stack returned by the topo_comp() function.}
}
\value{
A data.frame with the columns: `id`, `X93`, `Y93`, `maturing_doy`, `CSI_adj` and `MDI`.
}
\description{
This function calculates the maturing date (`maturing`) of bark beetle larvae after swarming (`swarming`), weighted by the topographic modulators CSIadj and MDI.
}
\examples{
\dontrun{
 maturing_table <- maturing(drias_data, swarming_table, topography)
}
}
