% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{ppc}
\alias{ppc}
\title{ppc}
\usage{
ppc(drias_table, topography)
}
\arguments{
\item{drias_table}{The DRIAS table processed by the phloem_rm() function.}

\item{topography}{The raster stack returned by the topo_comp() function.}
}
\value{
The updated DRIAS table with additional columns `cshd` containing the cast-shawoding factor, and `pp_cshd` containing the duration of the photoperiod for each day moderated by that factor.
}
\description{
This function calculates the photoperiod (time of incident light between sunrise and sunset) according to the latitude, doy and surrounding topography for the cast shadows.
}
\examples{
\dontrun{
 drias_table <- ppc(drias_table, topography)
}
}
