% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{swarming}
\alias{swarming}
\title{swarming}
\usage{
swarming(drias_table, awakening_table, topography)
}
\arguments{
\item{drias_table}{The DRIAS table processed by the phloem_rm() function.}

\item{awakening_table}{The data.frame returned by the awakening() function.}

\item{topography}{The raster stack returned by the topo_comp() function.}
}
\value{
A data.frame with the columns: `id`, `X93`, `Y93`, `awakening_doy`, and `swarming_doy`.
}
\description{
This function calculates the swarming day (`swarming`) of bark beetles after their awakening (`awakening`).
}
\examples{
\dontrun{
 swarming_table <- swarming(drias_data, awakening_table, topography)
}
}
