% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{awakening}
\alias{awakening}
\title{awakening}
\usage{
awakening(drias_table, topography)
}
\arguments{
\item{drias_table}{The DRIAS table processed by the phloem_rm() function.}

\item{topography}{The raster stack returned by the data_fetcher() function.}
}
\value{
A data.frame with the columns: `id`, `X93`, `Y93`, `awakening_doy`, and `CSI_adj` (adjusted CSI index).
}
\description{
This function calculates the awakening day (`awakening`) of adult bark beetles, weighted by the topographic modulator CSI_adj.
}
\examples{
\dontrun{
 awakening_table <- awakening(drias_data, topography)
}
}
