% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{drias_reader}
\alias{drias_reader}
\title{drias_reader}
\usage{
drias_reader(drias_txt_path)
}
\arguments{
\item{drias_txt_path}{Le chemin du fichier texte contenant les données climatiques DRIAS. Les données doivent être séparées par `,`, la date doit être au format JJ/MM/AAAA et elles doivent contenir tmin (°C), tmax (°C), tmean (°C), pr_tot (mm), spec_hum (kg/kg), vis_solrad (W/m2), ir_solrad (W/m2) et wind (m2/s).}
}
\value{
Un data.frame contenant les colonnes: `id`, `X93`, `Y93`, `date`, `doy`, `tmin`, `tmax`, `tmean`, `pr_tot`, `spec_hum`, `vis_solrad`, `ir_solrad`, `wind`.
}
\description{
Cette fonction lit un fichier texte contenant des données climatiques issues de DRIAS et les formate en un dataframe compatible avec le pipeline MPSS.
}
\examples{
\dontrun{
 drias_table <- drias_reader("Chablais_2030.txt")
}
}
