% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{hsc}
\alias{hsc}
\title{hsc}
\usage{
hsc(drias_table, topography)
}
\arguments{
\item{drias_table}{The DRIAS table processed by the ppc() function.}

\item{topography}{The raster stack returned by the topo_comp() function.}
}
\value{
The updated DRIAS table with additional column `hsi` containing the water stress intensity index for each day.
}
\description{
This function calculates the hydraulic stress index of spruce forests according to the water balance P-ETP and the AWC_max (Available Water Content) map.
}
\examples{
\dontrun{
 drias_table <- hsc(drias_table, topography)
}
}
