% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{kpi}
\alias{kpi}
\title{kpi}
\usage{
kpi(awakening_table, swarming_table, maturing_table, topography)
}
\arguments{
\item{awakening_table}{Le data.frame renvoyé par la fonction awakening().}

\item{swarming_table}{Le data.frame renvoyé par la fonction swarming().}

\item{maturing_table}{Le data.frame renvoyé par la fonction maturing().}

\item{topography}{Le stack raster renvoyé par la fonction data_fetcher().}
}
\value{
Un raster stack contenant les indicateurs phénologiques interpolés (`awakening_doy`, `swarming_doy`, `maturing_doy`), restreints aux pessières pures.
}
\description{
Cette fonction spatialise les indicateurs phénologiques (`awakening_doy`, `swarming_doy`, `maturing_doy`) via une IDW ou krigeage ordinaire à modèle exponentiel, en fonction du nombre de points DRIAS présents dans la zone d'étude.
}
\examples{
\dontrun{
 pheno_ind <- kpi(awakening_table, swarming_table, maturing_table, topography)
}
}
