% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{maturing}
\alias{maturing}
\title{maturing}
\usage{
maturing(drias_table, swarming_table, topography)
}
\arguments{
\item{drias_table}{La table DRIAS traitée par la fonction phloem_rm().}

\item{swarming_table}{Le data.frame renvoyé par la fonction swarming().}

\item{topography}{Le stack raster renvoyé par la fonction data_fetcher().}
}
\value{
Un data.frame avec les colonnes: `id` (identifiant des stations), `X93`, `Y93` (coordonnées), `maturing_doy` (jour de maturation des larves), `CSI_adj` (indice CSI ajusté), et `MDI` (indice de développement microclimatique).
}
\description{
Cette fonction calcule la date de maturation (`maturing`) des larves de scolytes après le vol (`swarming`), pondéré par les modulateurs topographiques CSIadj et MDI.
}
\examples{
\dontrun{
 maturing_table <- maturing(drias_data, swarming_table, topography)
}
}
