% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{phloem_rm}
\alias{phloem_rm}
\title{phloem_rm}
\usage{
phloem_rm(drias_table)
}
\arguments{
\item{drias_table}{La table DRIAS traitée par la fonction drias_reader().}
}
\value{
La table DRIAS mise à jour avec des colonnes supplémentaires `tmin_phloem`, `tmax_phloem` et `tmean_phloem` contenant les températures sous-phloème.
}
\description{
Cette fonction utilise un modèle radiatif non-linéaire contraint calibré sur l'épicéa pour calculer les températures sous phloème régissant le développement du scolyte.
}
\examples{
\dontrun{
 drias_table <- phloem_rm(drias_table)
}
}
