% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{phloem_rm}
\alias{phloem_rm}
\title{phloem_rm}
\usage{
phloem_rm(drias_table)
}
\arguments{
\item{drias_table}{The DRIAS table processed either by the drias_reader() or the drias_fetcher() function.}
}
\value{
The updated DRIAS table with additional columns `tmin_phloem`, `tmax_phloem`, and `tmean_phloem` containing the temperatures beneath the phloem.
}
\description{
This function uses a constrained non-linear radiative model calibrated for spruce to calculate the temperatures beneath the phloem regulating bark beetle development.
}
\examples{
\dontrun{
 drias_table <- phloem_rm(drias_table)
}
}
