% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{pipeline}
\alias{pipeline}
\title{pipeline}
\usage{
pipeline(topography, drias_table, return_tables = FALSE)
}
\arguments{
\item{topography}{The raster stack returned by the topo_comp() function.}

\item{drias_table}{The DRIAS table processed either by the drias_reader() or the drias_fetcher() function.}

\item{return_tables}{Whether the intermediate phenological tables (`awakening_table`, `swarming_table` and `maturing_table`) should be returned as ouputs of the pipeline or not.}
}
\value{
A list of the raster stack containing the spatialized phenological indicators (`awakening_doy`, `swarming_doy`, `maturing_doy`), the attack risk (`Rpheno`), and the maximum number of generations per year (`max_gen`), and the intermediate phenological tables if `return_tables = TRUE`.
}
\description{
Runs the entire B2SPM pipeline, from computing topography cast shadowing to spatializing the attack risk.
}
\examples{
\dontrun{
 results <- pipeline(topography, drias_table)
}
}
