% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{pipeline}
\alias{pipeline}
\title{pipeline}
\usage{
pipeline(drias_txt_path, dem)
}
\arguments{
\item{drias_txt_path}{Chemin vers le fichier DRIAS contenant les données météorologiques journalières.}

\item{dem}{Un raster (de préférence un SpatRaster en EPSG:2154) d'altitude couvrant l'entièreté de la zone d'étude + un buffer de 5px.}
}
\value{
Un stack raster contenant les indicateurs phénologiques spatialisés (`awakening_doy`, `swarming_doy`, `maturing_doy`), le risque d'attaque (`Rpheno`) et le nombre de générations maximum par année (`max_gen`).
}
\description{
Exécute l'ensemble du pipeline MPSS, de la lecture des données DRIAS à la spatialisation du risque d'attaque et du nombre d'attaques par an.
}
\examples{
\dontrun{
 dem <- terra::rast("dem_roi.tif")
 results <- pipeline("drias.txt", bbox)
}
}
