% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{pipeline}
\alias{pipeline}
\title{pipeline}
\usage{
pipeline(drias_txt_path, dem)
}
\arguments{
\item{drias_txt_path}{Path to the DRIAS file containing daily meteorological data.}

\item{dem}{A raster (preferably a SpatRaster in EPSG:2154) representing elevation, covering the entire study area plus a 5-pixel buffer.}
}
\value{
A raster stack containing the spatialized phenological indicators (`awakening_doy`, `swarming_doy`, `maturing_doy`), the attack risk (`Rpheno`), and the maximum number of generations per year (`max_gen`).
}
\description{
Runs the entire B2SPM pipeline, from reading DRIAS data to spatializing the attack risk and number of attacks per year.
}
\examples{
\dontrun{
 dem <- terra::rast("dem_roi.tif")
 results <- pipeline("drias.txt", bbox)
}
}
