% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{rpc}
\alias{rpc}
\title{rpc}
\usage{
rpc(pheno_ind)
}
\arguments{
\item{pheno_ind}{Le stack raster renvoyé par la fonction kpi().}
}
\value{
Un raster de probabilité (`Rpheno`), compris entre 0 et 1, indiquant le risque épidémique global de développement d'une génération de scolytes et donc d'une attaque.
}
\description{
Cette fonction calcule l'indicateur global de risque épidémique (`Rpheno`), mesurant la probabilité de succès du développement d'une génération de scolytes en fonction des indicateurs `awakening_doy`, `swarming_doy` et `maturing_doy`, et donc du risque d'attaque.
}
\examples{
\dontrun{
 rpheno <- rpc(pheno_ind)
}
}
