% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{swarming}
\alias{swarming}
\title{swarming}
\usage{
swarming(drias_table, awakening_table, topography)
}
\arguments{
\item{drias_table}{La table DRIAS traitée par la fonction phloem_rm().}

\item{awakening_table}{Le data.frame renvoyé par la fonction awakening().}

\item{topography}{Le stack raster renvoyé par la fonction data_fetcher().}
}
\value{
Un data.frame avec les colonnes: `id` (identifiant des stations), `X93`, `Y93` (coordonnées), `awakening_doy` (jour de réveil) et `swarming_doy` (jour du vol).
}
\description{
Cette fonction calcule le jour de vol (`swarming`) des scolytes après leur réveil (`awakening`).
}
\examples{
\dontrun{
 swarming_table <- swarming(drias_data, awakening_table, topography)
}
}
