% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.r
\name{topo_comp}
\alias{topo_comp}
\title{topo_comp}
\usage{
topo_comp(dem)
}
\arguments{
\item{dem}{Un raster (de préférence un SpatRaster en EPSG:2154) d'altitude couvrant l'entièreté de la zone d'étude + un buffer de 5px.}
}
\value{
Un stack raster contenant le masque des pessières et chaque raster topographique dans une bande séparée.
}
\description{
Cette fonction charge le masque des pessières et calcule les données topographiques sur la zone d'étude.
}
\examples{
\dontrun{
 dem <- terra::rast("dem_roi.tif")
 topography <- topo_comp(dem)
}
}
